refactor(activerecord): address Copilot review on PR #557
- Document LineDelta insertedAtLine=-1 sentinel for prepended blocks
- Extract shared resolveAssociationTarget helper used by synthesize +
  auto-import so declared types and auto-imports can't drift
- Update collectBaseDescendants JSDoc to describe WalkerResult
- Add explicit remapLine(2, deltas) === 0 assertion in prepend test
- Pass 1 of createTrailsProgram now uses a plain ts.CompilerHost so
  the text transform only runs once per build
